загрузка крупных файлов, нескольких файлов есть

+++ backend/app/models/message.py
@@ -1,7 +1,7 @@
 # app/models/message.py
 from enum import Enum as PyEnum
 
-from sqlalchemy import Boolean, Enum, ForeignKey, Integer, String, Text
+from sqlalchemy import BigInteger, Boolean, Enum, ForeignKey, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.models.base import BaseModel
@@ -41,7 +41,7 @@ class Message(BaseModel):
         index=True,
     )
     file_url: Mapped[str | None] = mapped_column(String(500), nullable=True)
-    file_size: Mapped[int | None] = mapped_column(Integer, nullable=True)  # байты
+    file_size: Mapped[int | None] = mapped_column(BigInteger, nullable=True)  # байты
     duration: Mapped[int | None] = mapped_column(Integer, nullable=True)  # секунды
 
     filename: Mapped[str | None] = mapped_column(String(255), nullable=True)
